refactor(sse-dashboard): split monolithic main.rs into modules
