+++ examples/sse-dashboard/src/config.rs
@@ -1,0 +1,8 @@
+use modo::AppConfig;
+use serde::Deserialize;
+
+#[derive(Default, Deserialize)]
+pub(crate) struct Config {
+    #[serde(flatten)]
+    pub(crate) core: AppConfig,
+}
